mram: Abstract the mram interface better
Remove all external use of mram devices directly.  Do everything by mram
number (for the status interfaces in ConsoleTask) or by addressing the
whole mram.

Move the MRAM reading and writing into the MRAM code and remove it from
the nonvol code.

This is getting ready to partition the mram, and provides a better
abstraction.

Signed-off-by: Corey Minyard <minyard@acm.org>