filesystem: Fix parameters to support a small filesystem
The setup for filesystems was not working with a 64K memory space, it
was taking up most of the space with inodes.  Decrease the space used by
inodes by:

  * Reducing the number of inodes to 50
  * Changing the indirect pointers per inode to 14

This will limit the number of files+directories to < 50, depending on
file size.  If the size of the file <= 512, then it will use a single
inode (all direct inode pointers), but if it's larger than that, it will
use indirect inode references and those will use up blocks.

In addition, reduce the block size to 128 to allow more efficient memory
usage.

With this, an empty filesystem takes has 51584 free bytes.  Before it
was ~12000.

Signed-off-by: Corey Minyard <minyard@acm.org>