Authenticate: Rework key handling
Resize the key to 32 bytes everywhere and use sizeof or AUTH_KEY_SIZE
for the size, get rid of the hard-coded numbers.

Move the hmac_sha_key and key handling functions to keyfile.c.  Add
a function to load the key from MRAM.

Rework the key storage to store a 32-byte key and to use the keyfile
checksum function.

Signed-off-by: Corey Minyard <minyard@acm.org>