Clean up and fix TxTask
Clean up the formatting to make it consistent.

Modify it to power on the PA when it starts transmitting, and power
it off when it stops.  I'm not sure how long it takes to power up
the PA, a delay may have to be added after the power on.  The spec
sheet of the TQP7M9106 doesn't say anything about it.

Also, don't turn of the PA between each packet if multiple
packets are pending.

Signed-off-by: Corey Minyard <corey@minyard.net>