Clean up ax5043_access
Make coding style consistent.

Remove open-coded ax5043 operations from the console and rx tasks and
put them in ax5043_access.

Remove some unused functions and pull as much private as possible.

Rework number of AX5043s to be consistent with the target device
and not always 5.

Fix compilation issues in the SPI driver.

Signed-off-by: Corey Minyard <corey@minyard.net>