rtc: Clean things up and fix power off handling
Lots of code cleanups.

Change the power management to always power the chip off of Vbat.  If
you don't, it won't properly survive a power outage.  That's not how
I understand it should work, but this is the only way I could make it
work and it's an ok compromise.

Signed-off-by: Corey Minyard <corey@minyard.net>