Add handling for invalid time
If the time is not valid in the RTC, do not allow the time to be
fetched from the RTC and make sure the time stays invalid in the
RTC by disabling the oscillator.  If the time is set, then enable
the oscillator and set the time as valid.

Add a bool to say if the unix time is valid.  Just print out
a lot if it's not valid, no handling for that yet.

Signed-off-by: Corey Minyard <corey@minyard.net>